Refactor(User): User 패키지 생성
유저 관련 클래스를 그룹화하기 위해 user 패키지를 만들었습니다.
패키지를 새로 생성해서 테스트 코드에 import문을 주가했습니다.
