+++ src/main/java/kr/codesqaud/cafe/user/User.java
@@ -1,4 +1,4 @@
-package kr.codesqaud.cafe;
+package kr.codesqaud.cafe.user;
 
 public class User {
     private String userId;
